Added interfaces to hide the implementation of public setters.
