+++ src/main/java/org/jboss/jdf/stacks/model/Archetype.java
@@ -1,151 +1,22 @@
-/*
- * JBoss, Home of Professional Open Source
- * Copyright 2013, Red Hat, Inc. and/or its affiliates, and individual
- * contributors by the @authors tag. See the copyright.txt in the
- * distribution for a full listing of individual contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 package org.jboss.jdf.stacks.model;
 
-public class Archetype {
+public interface Archetype {
 
-    private String id;
-
-    private String name;
-
-    private String description;
-
-    private String groupId;
-
-    private String artifactId;
-
-    private String recommendedVersion;
+    public String getId() ;
     
-    private String repositoryURL;
-
-    private Archetype blank;
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public String getGroupId() {
-        return groupId;
-    }
-
-    public void setGroupId(String groupId) {
-        this.groupId = groupId;
-    }
-
-    public String getArtifactId() {
-        return artifactId;
-    }
-
-    public void setArtifactId(String artifactId) {
-        this.artifactId = artifactId;
-    }
-
-    public String getRecommendedVersion() {
-        return recommendedVersion;
-    }
-
-    public void setRecommendedVersion(String recommendedVersion) {
-        this.recommendedVersion = recommendedVersion;
-    }
+    public String getName() ;
     
-    public String getRepositoryURL() {
-        return repositoryURL;
-    }
+    public String getDescription() ;
+
+    public String getGroupId() ;
+
+    public String getArtifactId() ;
     
-    public void setRepositoryURL(String repositoryURL) {
-        this.repositoryURL = repositoryURL;
-    }
+    public String getRecommendedVersion();
+        
+    public String getRepositoryURL() ;
+    
+    public Archetype getBlank() ;
 
-    public Archetype getBlank() {
-        return blank;
-    }
-
-    public void setBlank(Archetype blank) {
-        this.blank = blank;
-    }
-
-    /*
-     * (non-Javadoc)
-     * 
-     * @see java.lang.Object#hashCode()
-     */
-    @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = 1;
-        result = prime * result + ((id == null) ? 0 : id.hashCode());
-        return result;
-    }
-
-    /*
-     * (non-Javadoc)
-     * 
-     * @see java.lang.Object#equals(java.lang.Object)
-     */
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj) {
-            return true;
-        }
-        if (obj == null) {
-            return false;
-        }
-        if (!(obj instanceof Archetype)) {
-            return false;
-        }
-        Archetype other = (Archetype) obj;
-        if (id == null) {
-            if (other.id != null) {
-                return false;
-            }
-        } else if (!id.equals(other.id)) {
-            return false;
-        }
-        return true;
-    }
-
-    /*
-     * (non-Javadoc)
-     * 
-     * @see java.lang.Object#toString()
-     */
-    @Override
-    public String toString() {
-        return "Archetype [id=" + id + "]";
-    }
-
+    	
 }
